requirements.txt
flask==3.0.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
llama-index-core==0.10.10
llama-index-llms-openai==0.1.6
llama-index-llms-gemini==0.1.5
google-generativeai==0.3.2
plotly==5.19.0
python-dotenv==1.0.1
markupsafe==2.1.5
matplotlib==3.8.3
seaborn==0.13.2
xgboost==2.0.3